backend/pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "swmra-backend"
version = "0.1.0"
description = "Smart Waste Management & Recycling Assistant backend"
authors = [{name = "SWMRA", email = "dev@swmra.local"}]
requires-python = ">=3.11"
dependencies = [
  "fastapi~=0.111",
  "uvicorn[standard]~=0.30",
  "sqlmodel~=0.0.16",
  "aiosqlite~=0.20",
  "pydantic-settings~=2.4",
  "PyJWT~=2.8",
  "passlib[bcrypt]~=1.7",
  "python-multipart~=0.0.9",
  "apscheduler~=3.10",
  "email-validator~=2.1",
  "httpx~=0.27",
  "aiosmtplib~=2.0",
  "pillow~=10.4",
  "aiofiles~=24.1",
  "python-slugify~=8.0"
]

[project.optional-dependencies]
test = [
  "pytest~=8.2",
  "pytest-asyncio~=0.23",
  "httpx~=0.27",
  "anyio~=4.4",
  "ruff~=0.5",
  "black~=24.4",
  "isort~=5.13"
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-q"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

